App for running background subtraction.
Provides a QT gui for testing different background subtraction
parameters. Results are still pretty bad. Adds a "dataset"
directory to the root of the repo as a standard place for putting
downloaded datasets. Adds a vscode config that allows running of the
background subtraction app via a launch configuration, and a .env file
to set the PYTHONPATH to find the camera_control python package.