Rough implementation of pan-tilt-zoom to point
When the user clicks on a point in the live camera image, the camera
will pan and tilt to look at that point, and zoom in a bit. It's not
very accurate yet. The camera seems to always zoom at the same speed
regardless of what it is told, and ignores zero values sent for pan and
tilt speeds if the camera is already moving (it just keeps moving).