Use ruff for python formatting
We were using the black formatter, but it was configured incorrectly
and so it wasn't actually running. Now we're using ruff, since it
also provides an import formatter. You'll need to install the ruff
extension for vscode (you should be prompted to do so) To format code,
press ctrl-shift-I. To organize imports, press alt-shift-O.