Cleanup inference script
Use Dataset class directly instead of get_data_loader to avoid
batching (which is not needed for inference). Store inference results
directly in the pandas dataframe instead of having extra lists. Handle
multiple outputs smoothly for both inference and plotting so that we
don't need to special case when use_water_line is True.