Treat edge of board as a wall for jump checking
Also fixes the comparison logic that makes sure that the moves allowed
by the Quoridor class are the same as the moves expected by the test.
There were two problems before. When the two lists of moves were
different lengths, zip() was ignoring the elements at the end of
the longer array. Also the assert was on a list of truth values, and
would always evaluate as true as long as the lists weren't of length zero.
  To fix this, we now convert the lists of moves to 2d numpy arrays of
size (num_points, 2) and then assert equality with
np.testing.assert_equal.