Adding multiple new capabilities
New rotation utils with functions to rotate actions masks, actions, board and walls
Add a way for dumping a json_config to print full args on the screen
Encoded DExp play parameters in the file name so agent configures itself when reading the file. Hacky but good for know until we implement some level of config and tracking of training rounds.
Move epsilon decay to end of game instead of per-transition
Allows an extra argument after the classname during play to pass some basic info. Dexp-191 will pass 191 to the constructor of DExp. It is responsibility of the agent to interpret it. Note that agent will use those values to load the proper model.
Allows the use of a multiplier for final results, configurable in the trainable agent
DExp has support for: rotation, split of board for P1 and P2, inclusion of whether it is agent's turn
DExp has a partial implementation to incorporate opponents transitions (not finished, it is only enabled by param). Hook methods are already added to trainable agent