Multiple minor fixes (#165)
Avoids loading training params when loading a model. During play those should not be used, and if training, new params should be passed.
Fix negative rewards that was a TODO and was broken from previous PR.
Fix import that is using src.deep_quoridor to match current style being used.
Fix training status renderer to skip agents if they are not in training mode.