Fixing issues after new implementation (#181)
Fix two issues in the new quoridor impl:

Making env.observe compatible with past environment. New env was using 1 and 2 on the board for player 1 and player 2 but previous observation was using 1 for the player being observed, and 2 for the opponent
-- Fix construction of board based on observation to match fixed observation
Reward passed for the handle opponent was using the agent reward instead of the opponent.
Some other changes included here:
Smaller NN - to ensure I could reproduce 100% win rate against greedy
Making LR configurable as parameter
Adding agents for text board renderer to know which player is which
Nits